% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_many_vk_models.R
\name{fit_many_vk_models}
\alias{fit_many_vk_models}
\title{For a viral kinetic dataset in nlmixr format
This assesses how parsimoniously several viral kinetic models fit the data
Returned are AIC values. As parameter identifiability is also crucial the max,
mean, sd of the RSE\% for populations parameter estiamates are returned also.}
\usage{
fit_many_vk_models(vl_data = data2_cmv_untreated, num_fit_attempts = 5)
}
\arguments{
\item{vl_data}{nlmixr formatted viral load data with or without drug admininstrations}

\item{num_fit_attempts}{for each model, the number or attempts at fitting with random init pop parameter vectors}
}
\value{
Table outlining AIC and RSE comparisons of the viral load models
}
\description{
For a viral kinetic dataset in nlmixr format
This assesses how parsimoniously several viral kinetic models fit the data
Returned are AIC values. As parameter identifiability is also crucial the max,
mean, sd of the RSE\% for populations parameter estiamates are returned also.
}
