% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_model.r
\name{handle_messy_long_compute}
\alias{handle_messy_long_compute}
\title{handles long computes and time outs nicely}
\usage{
handle_messy_long_compute(model_string, data, time_limit = 10)
}
\arguments{
\item{model_string}{nlmixr compatible ode function as string}

\item{data}{a nlmixr compatible dataset}

\item{time_limit}{the transient time that passes before the command is killed.}
}
\value{
String of "error" or "pass"
}
\description{
For models requireing a long burn in, and where convergence is slow we can get many lsoda errors.
This function simply kills the argument nicely if it's taken longer than n seconds.
This is helpful when initial parameter conditions were poor, and enables the brute force approach.
}
