% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m13_SARSTCLMwImmune}
\alias{m13_SARSTCLMwImmune}
\title{m13_SARSTCLMwImmune   immune  recovers target cells}
\usage{
m13_SARSTCLMwImmune(fix_E0 = T, ...)
}
\value{
VK nlmixr function string
}
\description{
m13_SARSTCLMwImmune   immune  recovers target cells
}
