% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m14_SARSTCLMwImmune}
\alias{m14_SARSTCLMwImmune}
\title{m14_SARSTCLMwImmune immune blocks cell infections}
\usage{
m14_SARSTCLMwImmune(fix_E0 = T, ...)
}
\value{
VK nlmixr function string
}
\description{
m14_SARSTCLMwImmune immune blocks cell infections
}
