% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m15_SARSTCLMwImmune}
\alias{m15_SARSTCLMwImmune}
\title{m15_SARSTCLMwImmune immune increases viral clearance}
\usage{
m15_SARSTCLMwImmune(fix_E0 = T, ...)
}
\value{
VK nlmixr function string
}
\description{
m15_SARSTCLMwImmune immune increases viral clearance
}
