% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m16_SARSTCLMwImmune}
\alias{m16_SARSTCLMwImmune}
\title{m16_SARSTCLMwImmune immune promotes cytotoxicity}
\usage{
m16_SARSTCLMwImmune(fix_E0 = T, ...)
}
\value{
VK nlmixr function string
}
\description{
m16_SARSTCLMwImmune immune promotes cytotoxicity
}
