% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m17_hepc1}
\alias{m17_hepc1}
\title{m17_hepc1
doi: 10.1128/AAC.02611-14}
\usage{
m17_hepc1(fix_E0 = T, ...)
}
\value{
VK nlmixr function string
}
\description{
m17_hepc1
doi: 10.1128/AAC.02611-14
}
