% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m1_MwTCLESIC}
\alias{m1_MwTCLESIC}
\title{m1_MwTCLESIC}
\usage{
m1_MwTCLESIC(fix_dS = F, fix_V_v = F, fix_dI = F, fix_dE = F, fix_gamma = F)
}
\value{
VK nlmixr function string
}
\description{
m1_MwTCLESIC
}
