% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m2_MwTCLEDIS}
\alias{m2_MwTCLEDIS}
\title{m2_MwTCLEDIS}
\usage{
m2_MwTCLEDIS(fix_dS = F, fix_E_v = F, fix_dI = F, fix_dE = F, fix_gamma = F)
}
\value{
VK nlmixr function string
}
\description{
m2_MwTCLEDIS
}
