% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_kinetic_models.R
\name{m6_EDICwTCL}
\alias{m6_EDICwTCL}
\title{m6_EDICwTCL}
\usage{
m6_EDICwTCL(fix_dI = F, fix_dE = F, fix_E_v = F, fix_V_v = F, ...)
}
\value{
VK nlmixr function string
}
\description{
m6_EDICwTCL
}
