% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ode_string_to_function}
\alias{ode_string_to_function}
\title{load function to memory
definitely no security risks here..  //todo}
\usage{
ode_string_to_function(ode_function_string)
}
\arguments{
\item{ode_function_string}{string defining nlmixr compatible ode}
}
\value{
String of "error" or "pass"
}
\description{
Simply handles whether the model can compile via nlmixr
}
